pyproject.toml
[tool.poetry]
name = "ifdash"
version = "0.1.0"
description = ""
authors = ["Thanathip Limna <tl@coe.psu.ac.th>"]

[tool.poetry.dependencies]
python="^3.12"
dash = {extras = ["diskcache"], version = "^2.14.0"}
dash-bootstrap-components = "^1.4.2"
rq = "^1.15.1"
flask-wtf = "^1.1.1"
flask-login = "^0.6.2"
httpx = "^0.24.1"
python-dotenv = "^1.0.0"
influxdb-client = "^1.37.0"
uptime-kuma-api = "^1.2.1"
flask-caching = "^2.1.0"
diskcache = "^5.6.3"
celery = "^5.3.4"
dash-leaflet = "^1.0.11"
beanie = "^1.25.0"
pydantic-settings = "^2.2.1"
nest-asyncio = "^1.6.0"
pandas = "^2.2.2"
authlib = "^1.3.0"
markdown = "^3.6"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
ifdash-dash = "ifdash.cmd.web:main"
ifdash-worker = "ifdash.cmd.worker:main"
ifdash-controller = "ifdash.cmd.controller:main"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
